fix(struct): make create_random as part of mocks
